Third step of integrating DBContext and Infrastructure and Application projects to Authorization
