+++ source/marketparticipant/Energinet.DataHub.MarketParticipant.Authorization.Infrastructure/Energinet.DataHub.MarketParticipant.Authorization.Infrastructure.csproj
@@ -16,4 +16,8 @@
     <ProjectReference Include="..\Energinet.DataHub.MarketParticipant.Authorization\Energinet.DataHub.MarketParticipant.Authorization.csproj" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="Model\" />
+  </ItemGroup>
+
 </Project>
