+++ source/marketparticipant/Energinet.DataHub.MarketParticipant.Authorization.Infrastructure/Persistence/EntityConfiguration/AuditedEntityTypeConfiguration.cs
@@ -13,22 +13,19 @@
 // limitations under the License.
 
 using System;
-using Energinet.DataHub.MarketParticipant.Authorization.Infrastructure.Domain;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
 
 namespace Energinet.DataHub.MarketParticipant.Authorization.Infrastructure.Persistence.EntityConfiguration;
 
 public abstract class AuditedEntityTypeConfiguration<TEntity> : IEntityTypeConfiguration<TEntity>
-    where TEntity : class, IAuditedEntity
+    where TEntity : class
 {
     public void Configure(EntityTypeBuilder<TEntity> builder)
     {
         ArgumentNullException.ThrowIfNull(builder, nameof(builder));
 
         builder.ToTable(t => t.IsTemporal());
-        builder.Property(e => e.Version).IsConcurrencyToken();
-        builder.Property(e => e.ChangedByIdentityId).IsRequired();
 
         ConfigureEntity(builder);
     }
