+++ source/marketparticipant/Energinet.DataHub.MarketParticipant.Authorization/Repositories/Model/ActorName.cs
@@ -12,6 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Energinet.DataHub.MarketParticipant.Authorization.Infrastructure.Model;
+namespace Energinet.DataHub.MarketParticipant.Authorization.Repositories.Model;
 
-public sealed record GridAreaCode(string Value);
+public sealed record ActorName(string Value);
