+++ source/marketparticipant/Energinet.DataHub.MarketParticipant.Authorization/Repositories/Model/GridAreaLink.cs
@@ -12,17 +12,24 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using System.ComponentModel;
+using System;
 
-namespace Energinet.DataHub.MarketParticipant.Authorization.Infrastructure.Model;
+namespace Energinet.DataHub.MarketParticipant.Authorization.Repositories.Model;
 
-public sealed record UnknownActorNumber : ActorNumber
+public sealed class GridAreaLink
 {
-    [Browsable(false)]
-    public UnknownActorNumber(string value)
-        : base(value)
+    public GridAreaLink(GridAreaId gridAreaId)
     {
+        Id = new GridAreaLinkId(Guid.Empty);
+        GridAreaId = gridAreaId;
     }
 
-    public override ActorNumberType Type => ActorNumberType.Unknown;
+    public GridAreaLink(GridAreaLinkId id, GridAreaId gridAreaId)
+    {
+        Id = id;
+        GridAreaId = gridAreaId;
+    }
+
+    public GridAreaLinkId Id { get; init; }
+    public GridAreaId GridAreaId { get; init; }
 }
