+++ source/marketparticipant/Energinet.DataHub.MarketParticipant.IntegrationTests/Repositories/Authorization/GridAreaRepositoryTests.cs
@@ -14,9 +14,8 @@
 
 using System;
 using System.Threading.Tasks;
-using Energinet.DataHub.MarketParticipant.Authorization.Infrastructure.Domain;
-using Energinet.DataHub.MarketParticipant.Authorization.Infrastructure.Model;
 using Energinet.DataHub.MarketParticipant.Authorization.Infrastructure.Persistence.Repositories;
+using Energinet.DataHub.MarketParticipant.Authorization.Repositories.Model;
 using Energinet.DataHub.MarketParticipant.IntegrationTests.Fixtures;
 using Xunit;
 using Xunit.Categories;
